api/customer-churn-api/requirements.txt
uvicorn>=0.11.3,<0.12.0
fastapi>=0.64.0,<1.0.0
python-multipart>=0.0.5,<0.1.0
pydantic>=1.8.1,<1.9.0
typing_extensions>=3.7.4,<3.8.0
loguru>=0.5.3,<0.6.0
# We will explain this in the course
slalom_classification_model==0.0.1